+++ exonum/src/runtime/rust/service.rs
@@ -25,15 +25,17 @@ use crate::{
     blockchain::config::InstanceInitParams,
     crypto::{Hash, PublicKey, SecretKey},
     helpers::{Height, ValidatorId},
-    messages::Verified,
     node::ApiSender,
     runtime::{
         dispatcher::{Action, Mailbox},
-        AnyTx, ArtifactId, CallInfo, ExecutionError, InstanceDescriptor, InstanceId, MethodId,
+        rust::GenericCall,
+        ArtifactId, ExecutionError, InstanceDescriptor, InstanceId, MethodId,
     },
 };
 
-use super::{api::ServiceApiBuilder, ArtifactProtobufSpec, BlockchainData, CallContext};
+use super::{
+    api::ServiceApiBuilder, ArtifactProtobufSpec, BlockchainData, CallContext, MethodDescriptor,
+};
 
 /// Describes how the service instance should dispatch specific method calls
 /// with consideration of the interface where the method belongs.
@@ -160,36 +162,6 @@ pub trait DefaultInstance: ServiceFactory {
     fn default_instance(&self) -> InstanceInitParams {
         self.artifact_id()
             .into_default_instance(Self::INSTANCE_ID, Self::INSTANCE_NAME)
-    }
-}
-
-/// Transaction specification for a specific service interface method.
-pub trait Transaction<Svc: ?Sized>: BinaryValue {
-    /// Identifier of the service interface required for the call.
-    #[doc(hidden)]
-    const INTERFACE_NAME: &'static str;
-    /// Identifier of the service method which executes the given transaction.
-    const METHOD_ID: MethodId;
-
-    /// Creates an unsigned service transaction from the value.
-    fn into_any_tx(self, instance_id: InstanceId) -> AnyTx {
-        AnyTx {
-            call_info: CallInfo {
-                instance_id,
-                method_id: Self::METHOD_ID,
-            },
-            arguments: self.into_bytes(),
-        }
-    }
-
-    /// Signs the value as a transaction with the specified instance identifier.
-    fn sign(
-        self,
-        service_id: InstanceId,
-        public_key: PublicKey,
-        secret_key: &SecretKey,
-    ) -> Verified<AnyTx> {
-        Verified::from_value(self.into_any_tx(service_id), public_key, secret_key)
     }
 }
 
@@ -345,26 +317,6 @@ impl<'a> Broadcaster<'a> {
         self.instance.as_ref()
     }
 
-    /// Signs and broadcasts a transaction to the other nodes in the network.
-    ///
-    /// The transaction is signed by the service keypair of the node. The same input transaction
-    /// will lead to the identical transaction being broadcast. If this is undesired, add a nonce
-    /// field to the input transaction (e.g., a `u64`) and change it between the calls.
-    ///
-    /// # Return value
-    ///
-    /// Returns the hash of the created transaction, or an error if the transaction cannot be
-    /// broadcast. An error means that the node is being shut down.
-    pub fn send<Svc: ?Sized, T>(self, tx: T) -> Result<Hash, Error>
-    where
-        T: Transaction<Svc>,
-    {
-        let (public_key, secret_key) = self.service_keypair.as_ref();
-        let msg = tx.sign(self.instance().id, *public_key, secret_key);
-        let tx_hash = msg.object_hash();
-        self.tx_sender.broadcast_transaction(msg).map(|()| tx_hash)
-    }
-
     /// Converts the broadcaster into the owned representation, which can be used to broadcast
     /// transactions asynchronously.
     pub fn into_owned(self) -> Broadcaster<'static> {
@@ -373,6 +325,28 @@ impl<'a> Broadcaster<'a> {
             service_keypair: Cow::Owned(self.service_keypair.into_owned()),
             tx_sender: Cow::Owned(self.tx_sender.into_owned()),
         }
+    }
+}
+
+/// Signs and broadcasts a transaction to the other nodes in the network.
+///
+/// The transaction is signed by the service keypair of the node. The same input transaction
+/// will lead to the identical transaction being broadcast. If this is undesired, add a nonce
+/// field to the input transaction (e.g., a `u64`) and change it between the calls.
+///
+/// # Return value
+///
+/// Returns the hash of the created transaction, or an error if the transaction cannot be
+/// broadcast. An error means that the node is being shut down.
+impl GenericCall<()> for Broadcaster<'_> {
+    type Output = Result<Hash, Error>;
+
+    fn generic_call(&self, _ctx: (), method: MethodDescriptor<'_>, args: Vec<u8>) -> Self::Output {
+        let msg = self
+            .service_keypair
+            .generic_call(self.instance().id, method, args);
+        let tx_hash = msg.object_hash();
+        self.tx_sender.broadcast_transaction(msg).map(|()| tx_hash)
     }
 }
 
@@ -409,19 +383,6 @@ impl Debug for AfterCommitContext<'_> {
             .field("instance", &self.broadcaster.instance)
             .finish()
     }
-}
-
-/// A service interface specification.
-pub trait Interface {
-    /// Fully qualified name of this interface.
-    const INTERFACE_NAME: &'static str;
-    /// Invokes the specified method handler of the service instance.
-    fn dispatch(
-        &self,
-        context: CallContext<'_>,
-        method: MethodId,
-        payload: &[u8],
-    ) -> Result<(), ExecutionError>;
 }
 
 fn is_supervisor(instance_id: InstanceId) -> bool {