+++ test-suite/testkit/tests/interfaces/main.rs
@@ -15,15 +15,15 @@
 use exonum::{
     crypto,
     messages::{AnyTx, Verified},
-    runtime::{rust::Transaction, CallInfo, DispatcherError, ErrorMatch, ExecutionError},
+    runtime::{CallInfo, DispatcherError, ErrorMatch, ExecutionError},
 };
 use exonum_testkit::{TestKit, TestKitBuilder};
 
 use crate::{
     error::Error,
     services::{
-        AnyCall, AnyCallService, DepositService, RecursiveCall, TxCreateWallet, TxIssue,
-        WalletService,
+        AnyCall, AnyCallService, CallAny, DepositInterface, DepositService, TxIssue,
+        WalletInterface, WalletService,
     },
 };
 
@@ -55,10 +55,7 @@ fn test_create_wallet_ok() {
 
     execute_transaction(
         &mut testkit,
-        TxCreateWallet {
-            name: "Alice".into(),
-        }
-        .sign(WalletService::ID, keypair.0, &keypair.1),
+        keypair.create_wallet(WalletService::ID, "Alice".into()),
     )
     .expect("Unable to create wallet");
 }
@@ -70,26 +67,18 @@ fn test_create_wallet_fallthrough_auth() {
 
     // Without fallthrough auth, the call should fail: `create_wallet` expects the caller
     // to be external, and it is a service.
-    let mut call = AnyCall::new(
-        CallInfo::new(WalletService::ID, 0),
-        TxCreateWallet {
-            name: "Alice".into(),
-        },
-    );
+    let mut call = AnyCall::new(CallInfo::new(WalletService::ID, 0), "Alice".to_owned());
     let err = execute_transaction(
         &mut testkit,
-        call.clone().sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_any(AnyCallService::ID, call.clone()),
     )
     .unwrap_err();
     assert_eq!(err, ErrorMatch::from_fail(&Error::WrongInterfaceCaller));
 
     // With fallthrough auth, the call should succeed.
     call.fallthrough_auth = true;
-    execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .expect("Cannot create wallet");
+    execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call))
+        .expect("Cannot create wallet");
 
     let snapshot = testkit.snapshot();
     assert_eq!(
@@ -109,20 +98,19 @@ fn test_deposit_ok() {
 
     execute_transaction(
         &mut testkit,
-        TxCreateWallet {
-            name: "Alice".into(),
-        }
-        .sign(WalletService::ID, keypair.0, &keypair.1),
+        keypair.create_wallet(WalletService::ID, "Alice".into()),
     )
     .expect("Unable to create wallet");
 
     execute_transaction(
         &mut testkit,
-        TxIssue {
-            to: keypair.0,
-            amount: 10_000,
-        }
-        .sign(DepositService::ID, keypair.0, &keypair.1),
+        keypair.deposit(
+            DepositService::ID,
+            TxIssue {
+                to: keypair.0,
+                amount: 10_000,
+            },
+        ),
     )
     .expect("Unable to deposit wallet");
 
@@ -149,7 +137,7 @@ fn test_deposit_ok() {
     call.fallthrough_auth = true;
     execute_transaction(
         &mut testkit,
-        call.clone().sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_any(AnyCallService::ID, call.clone()),
     )
     .expect("Unable to deposit more funds");
 
@@ -169,11 +157,8 @@ fn test_deposit_ok() {
         call = AnyCall::new(CallInfo::new(AnyCallService::ID, 0), call);
         call.fallthrough_auth = true; // Must be set to `true` in all calls!
     }
-    execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .expect("Unable to deposit funds with high indirection");
+    execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call))
+        .expect("Unable to deposit funds with high indirection");
 
     let snapshot = testkit.snapshot();
     assert_eq!(
@@ -193,10 +178,7 @@ fn test_deposit_invalid_auth() {
 
     execute_transaction(
         &mut testkit,
-        TxCreateWallet {
-            name: "Alice".into(),
-        }
-        .sign(WalletService::ID, keypair.0, &keypair.1),
+        keypair.create_wallet(WalletService::ID, "Alice".into()),
     )
     .expect("Unable to create wallet");
 
@@ -210,7 +192,7 @@ fn test_deposit_invalid_auth() {
     // Do not set fallthrough auth, as in the previous example.
     let err = execute_transaction(
         &mut testkit,
-        call.clone().sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_any(AnyCallService::ID, call.clone()),
     )
     .unwrap_err();
     assert_eq!(err, ErrorMatch::from_fail(&Error::UnauthorizedIssuer));
@@ -226,7 +208,7 @@ fn test_deposit_invalid_auth() {
     // for the deposit service.
     let err = execute_transaction(
         &mut testkit,
-        call.clone().sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_any(AnyCallService::ID, call.clone()),
     )
     .unwrap_err();
     assert_eq!(err, ErrorMatch::from_fail(&Error::UnauthorizedIssuer));
@@ -239,11 +221,13 @@ fn test_deposit_err_issue_without_wallet() {
 
     let err = execute_transaction(
         &mut testkit,
-        TxIssue {
-            to: keypair.0,
-            amount: 10_000,
-        }
-        .sign(DepositService::ID, keypair.0, &keypair.1),
+        keypair.deposit(
+            DepositService::ID,
+            TxIssue {
+                to: keypair.0,
+                amount: 10_000,
+            },
+        ),
     )
     .unwrap_err();
 
@@ -260,10 +244,7 @@ fn test_any_call_ok_deposit() {
 
     execute_transaction(
         &mut testkit,
-        TxCreateWallet {
-            name: "Alice".into(),
-        }
-        .sign(WalletService::ID, keypair.0, &keypair.1),
+        keypair.create_wallet(WalletService::ID, "Alice".into()),
     )
     .expect("Unable to create wallet");
 
@@ -275,11 +256,8 @@ fn test_any_call_ok_deposit() {
         },
     );
     call.fallthrough_auth = true;
-    execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .expect("Unable to deposit wallet");
+    execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call))
+        .expect("Unable to deposit wallet");
 
     let snapshot = testkit.snapshot();
     assert_eq!(
@@ -299,10 +277,7 @@ fn test_any_call_err_deposit_unauthorized() {
 
     execute_transaction(
         &mut testkit,
-        TxCreateWallet {
-            name: "Alice".into(),
-        }
-        .sign(WalletService::ID, keypair.0, &keypair.1),
+        keypair.create_wallet(WalletService::ID, "Alice".into()),
     )
     .expect("Unable to create wallet");
 
@@ -314,11 +289,8 @@ fn test_any_call_err_deposit_unauthorized() {
         },
     );
     call.interface_name = "IssueReceiver".to_owned();
-    let err = execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .unwrap_err();
+    let err =
+        execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call)).unwrap_err();
 
     assert_eq!(
         err,
@@ -330,12 +302,9 @@ fn test_any_call_err_deposit_unauthorized() {
 fn test_any_call_err_unknown_instance() {
     let mut testkit = testkit_with_interfaces();
     let keypair = crypto::gen_keypair();
-
-    let err = execute_transaction(
-        &mut testkit,
-        AnyCall::new(CallInfo::new(10_000, 0), ()).sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .unwrap_err();
+    let call = AnyCall::new(CallInfo::new(10_000, 0), ());
+    let err =
+        execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call)).unwrap_err();
 
     assert_eq!(
         err,
@@ -350,11 +319,8 @@ fn test_any_call_err_unknown_interface() {
 
     let mut call = AnyCall::new(CallInfo::new(WalletService::ID, 0), ());
     call.interface_name = "FooFace".to_owned();
-    let err = execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .unwrap_err();
+    let err =
+        execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call)).unwrap_err();
 
     assert_eq!(
         err,
@@ -375,11 +341,8 @@ fn test_any_call_err_unknown_method() {
         },
     );
     call.interface_name = "IssueReceiver".to_owned();
-    let err = execute_transaction(
-        &mut testkit,
-        call.sign(AnyCallService::ID, keypair.0, &keypair.1),
-    )
-    .unwrap_err();
+    let err =
+        execute_transaction(&mut testkit, keypair.call_any(AnyCallService::ID, call)).unwrap_err();
 
     assert_eq!(err, ErrorMatch::from_fail(&DispatcherError::NoSuchMethod));
 }
@@ -390,20 +353,17 @@ fn test_any_call_err_wrong_arg() {
     let keypair = crypto::gen_keypair();
 
     let inner_call = AnyCall::new(CallInfo::new(10_000, 0), ());
+    let outer_call = AnyCall::new(CallInfo::new(WalletService::ID, 0), inner_call);
     let err = execute_transaction(
         &mut testkit,
-        AnyCall::new(CallInfo::new(WalletService::ID, 0), inner_call).sign(
-            AnyCallService::ID,
-            keypair.0,
-            &keypair.1,
-        ),
+        keypair.call_any(AnyCallService::ID, outer_call),
     )
     .unwrap_err();
 
     assert_eq!(
         err,
         ErrorMatch::from_fail(&DispatcherError::MalformedArguments)
-            .with_description_containing("Utf8Error")
+            .with_description_containing("invalid utf-8 sequence")
     );
 }
 
@@ -414,13 +374,13 @@ fn test_any_call_panic_recursion_limit() {
 
     execute_transaction(
         &mut testkit,
-        RecursiveCall { depth: 256 }.sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_recursive(AnyCallService::ID, 256),
     )
     .expect("Call stack depth is enough");
 
     let err = execute_transaction(
         &mut testkit,
-        RecursiveCall { depth: 257 }.sign(AnyCallService::ID, keypair.0, &keypair.1),
+        keypair.call_recursive(AnyCallService::ID, 300),
     )
     .unwrap_err();